modules/lithium.pytools/tools/hotspot.py
"""
WiFi Hotspot Manager
A cross-platform tool for managing WiFi hotspots with support for Linux and Windows.

Features:
- Create and manage WiFi hotspots
- Support multiple authentication methods (WPA-PSK, WPA2, WPA3)
- Save and load hotspot profiles
- Monitor connected clients
- Async operations support
- Comprehensive logging

Author: Your Name
License: MIT
Version: 1.0.0
"""

from __future__ import annotations
import subprocess
import platform
import argparse
from typing import Optional, Literal, Dict, Any
from dataclasses import dataclass
from enum import Enum
import shutil
from pathlib import Path
import json
from datetime import datetime
import asyncio
from loguru import logger
import sys

# Configure loguru logger with custom format and file rotation
logger.remove()
logger.add(
    sys.stderr,
    format="<green>{time:YYYY-MM-DD HH:mm:ss}</green> | <level>{level: <8}</level> | <cyan>{name}</cyan>:<cyan>{function}</cyan>:<cyan>{line}</cyan> - <level>{message}</level>",
    level="INFO"
)
logger.add("hotspot.log", rotation="500 MB", retention="10 days")


class Platform(Enum):
    """
    Supported operating system platforms.
    Used to determine platform-specific implementations.
    """
    LINUX = "linux"
    WINDOWS = "windows"
    MACOS = "darwin"


class AuthType(Enum):
    """
    WiFi authentication types supported by the hotspot.

    Members:
        WPA_PSK: WPA Personal authentication
        WPA2: WPA2 authentication (more secure than WPA)
        WPA3: Latest WPA3 authentication (most secure)
    """
    WPA_PSK = "wpa-psk"
    WPA2 = "wpa2"
    WPA3 = "wpa3"


class EncryptionType(Enum):
    """
    Encryption methods for securing the WiFi connection.

    Members:
        AES: Advanced Encryption Standard (recommended)
        TKIP: Temporal Key Integrity Protocol (legacy)
    """
    AES = "aes"
    TKIP = "tkip"


@dataclass
class HotspotConfig:
    """
    Configuration settings for a WiFi hotspot.

    Attributes:
        name: SSID of the hotspot
        password: Security key for authentication
        authentication: Type of authentication (WPA/WPA2/WPA3)
        encryption: Type of encryption (AES/TKIP)
        channel: WiFi channel number (1-14)
        max_clients: Maximum number of concurrent connections
        band: WiFi frequency band (bg/a/ac)
        hidden: Whether to hide the SSID
    """
    name: str = "MyHotspot"
    password: Optional[str] = None
    authentication: AuthType = AuthType.WPA_PSK
    encryption: EncryptionType = EncryptionType.AES
    channel: int = 11
    max_clients: int = 10
    band: Literal["bg", "a", "ac"] = "bg"
    hidden: bool = False


class HotspotManager:
    """
    Core class for managing WiFi hotspots across different platforms.

    Handles hotspot creation, configuration, monitoring, and profile management.
    Supports both synchronous and asynchronous operations.
    """

    def __init__(self):
        """
        Initialize the HotspotManager with platform detection and config loading.
        Creates necessary directories if they don't exist.
        """
        logger.debug("Initializing HotspotManager")
        self.platform = Platform(platform.system().lower())
        self.config_path = Path.home() / ".hotspot"
        self.config_path.mkdir(exist_ok=True)
        self._load_config()

    def _load_config(self) -> None:
        """
        Load saved hotspot configurations from a JSON file.
        If the file does not exist, initialize an empty configuration dictionary.
        """
        logger.debug("Loading hotspot configurations")
        config_file = self.config_path / "config.json"
        if config_file.exists():
            with open(config_file) as f:
                self.saved_configs = json.load(f)
            logger.info("Hotspot configurations loaded successfully")
        else:
            self.saved_configs = {}
            logger.info(
                "No existing configurations found, starting with an empty configuration")

    def _save_config(self) -> None:
        """
        Save the current hotspot configurations to a JSON file.
        """
        logger.debug("Saving hotspot configurations")
        with open(self.config_path / "config.json", "w") as f:
            json.dump(self.saved_configs, f, indent=4)
        logger.info("Hotspot configurations saved successfully")

    async def _run_command_async(self, cmd: list[str]) -> tuple[str, str]:
        """
        Run a shell command asynchronously.

        Args:
            cmd: List of command arguments.

        Returns:
            A tuple containing the standard output and standard error of the command.

        Raises:
            Exception: If the command execution fails.
        """
        logger.debug(f"Running command asynchronously: {' '.join(cmd)}")
        try:
            process = await asyncio.create_subprocess_exec(
                *cmd,
                stdout=asyncio.subprocess.PIPE,
                stderr=asyncio.subprocess.PIPE
            )
            stdout, stderr = await process.communicate()
            logger.debug(f"Command output: {stdout.decode()}")
            if stderr:
                logger.error(f"Command error: {stderr.decode()}")
            return stdout.decode(), stderr.decode()
        except Exception as e:
            logger.error(f"Command execution failed: {e}")
            raise

    def _run_command(self, cmd: list[str]) -> str:
        """
        Run a shell command synchronously.

        Args:
            cmd: List of command arguments.

        Returns:
            The standard output of the command.

        Raises:
            subprocess.CalledProcessError: If the command execution fails.
        """
        logger.debug(f"Running command synchronously: {' '.join(cmd)}")
        try:
            result = subprocess.run(
                cmd,
                check=True,
                stdout=subprocess.PIPE,
                stderr=subprocess.PIPE,
                text=True
            )
            logger.debug(f"Command output: {result.stdout}")
            return result.stdout
        except subprocess.CalledProcessError as e:
            logger.error(f"Command failed: {e.stderr}")
            raise

    async def start(self, config: HotspotConfig) -> None:
        """
        Start a WiFi hotspot with the given configuration.

        Args:
            config: HotspotConfig object containing the hotspot settings.

        Raises:
            ValueError: If the password is not provided.
            NotImplementedError: If the platform is not supported.
        """
        logger.info(f"Starting hotspot with name: {config.name}")

        if config.password is None:
            logger.error("Password is required when starting a hotspot")
            raise ValueError("Password is required when starting a hotspot")

        if self.platform == Platform.LINUX:
            await self._start_linux(config)
        elif self.platform == Platform.WINDOWS:
            await self._start_windows(config)
        else:
            logger.error(f"Platform {self.platform} not supported")
            raise NotImplementedError(
                f"Platform {self.platform} not supported")

        logger.success(f"Hotspot {config.name} is now running")

    async def _start_linux(self, config: HotspotConfig) -> None:
        """
        Start a WiFi hotspot on a Linux platform.

        Args:
            config: HotspotConfig object containing the hotspot settings.
        """
        logger.debug(f"Starting Linux hotspot with configuration: {config}")
        cmd = [
            'nmcli', 'dev', 'wifi', 'hotspot',
            'ifname', 'wlan0',
            'ssid', config.name,
            'password', config.password
        ]
        await self._run_command_async(cmd)

        # Configure additional parameters
        await self._configure_linux_hotspot(config)

    async def _start_windows(self, config: HotspotConfig) -> None:
        """
        Start a WiFi hotspot on a Windows platform.

        Args:
            config: HotspotConfig object containing the hotspot settings.
        """
        logger.debug(f"Starting Windows hotspot with configuration: {config}")
        cmd = [
            'netsh', 'wlan', 'set', 'hostednetwork',
            'mode=allow',
            f'ssid={config.name}',
            f'key={config.password}'
        ]
        await self._run_command_async(cmd)

        # Start the hotspot
        await self._run_command_async(['netsh', 'wlan', 'start', 'hostednetwork'])

    async def stop(self) -> None:
        """
        Stop the running WiFi hotspot.
        """
        logger.info("Stopping hotspot")

        if self.platform == Platform.LINUX:
            await self._run_command_async(['nmcli', 'connection', 'down', 'Hotspot'])
        elif self.platform == Platform.WINDOWS:
            await self._run_command_async(['netsh', 'wlan', 'stop', 'hostednetwork'])

        logger.success("Hotspot has been stopped")

    async def status(self) -> Dict[str, Any]:
        """
        Get the current status of the WiFi hotspot.

        Returns:
            A dictionary containing the status and details of the hotspot.
        """
        logger.info("Checking hotspot status")
        if self.platform == Platform.LINUX:
            stdout, _ = await self._run_command_async(['nmcli', 'dev', 'status'])
            connected = 'connected' in stdout

            if connected:
                details, _ = await self._run_command_async(['nmcli', 'connection', 'show', 'Hotspot'])
                logger.debug(f"Hotspot status: running, details: {details}")
                return {
                    "status": "running",
                    "details": details
                }
        elif self.platform == Platform.WINDOWS:
            stdout, _ = await self._run_command_async(['netsh', 'wlan', 'show', 'hostednetwork'])
            status = "running" if "Started" in stdout else "stopped"
            logger.debug(f"Hotspot status: {status}, details: {stdout}")
            return {
                "status": status,
                "details": stdout
            }

        logger.debug("Hotspot status: stopped")
        return {"status": "stopped"}

    async def list_clients(self) -> list[Dict[str, str]]:
        """
        List the clients connected to the WiFi hotspot.

        Returns:
            A list of dictionaries containing client information.
        """
        logger.info("Listing connected clients")
        clients = []

        if self.platform == Platform.LINUX:
            stdout, _ = await self._run_command_async(['iw', 'dev', 'wlan0', 'station', 'dump'])
            # Parse the output to get client information
            logger.debug(f"Connected clients (Linux): {stdout}")
            # TODO: Implement parsing logic

        elif self.platform == Platform.WINDOWS:
            stdout, _ = await self._run_command_async(['netsh', 'wlan', 'show', 'hostednetwork'])
            # Parse the output to get client information
            logger.debug(f"Connected clients (Windows): {stdout}")
            # TODO: Implement parsing logic

        return clients

    def save_profile(self, name: str, config: HotspotConfig) -> None:
        """
        Save a hotspot configuration profile.

        Args:
            name: Name of the profile.
            config: HotspotConfig object containing the hotspot settings.
        """
        logger.info(f"Saving profile: {name}")
        self.saved_configs[name] = {
            "name": config.name,
            "authentication": config.authentication.value,
            "encryption": config.encryption.value,
            "channel": config.channel,
            "max_clients": config.max_clients,
            "band": config.band,
            "hidden": config.hidden
        }
        self._save_config()
        logger.info(f"Profile {name} has been saved")

    def load_profile(self, name: str) -> HotspotConfig:
        """
        Load a hotspot configuration profile.

        Args:
            name: Name of the profile.

        Returns:
            A HotspotConfig object containing the loaded settings.

        Raises:
            ValueError: If the profile is not found.
        """
        logger.info(f"Loading profile: {name}")
        if name not in self.saved_configs:
            logger.error(f"Profile {name} not found")
            raise ValueError(f"Profile {name} not found")

        config_dict = self.saved_configs[name]
        logger.debug(
            f"Profile {name} loaded with configuration: {config_dict}")
        return HotspotConfig(
            name=config_dict["name"],
            authentication=AuthType(config_dict["authentication"]),
            encryption=EncryptionType(config_dict["encryption"]),
            channel=config_dict["channel"],
            max_clients=config_dict["max_clients"],
            band=config_dict["band"],
            hidden=config_dict["hidden"]
        )


async def main():
    """
    Main function to parse command-line arguments and perform actions.
    """
    parser = argparse.ArgumentParser(
        description='Advanced WiFi Hotspot Manager')
    parser.add_argument('action', choices=['start', 'stop', 'status', 'list-clients', 'save-profile', 'load-profile'],
                        help='Action to perform')
    parser.add_argument('--name', default='MyHotspot', help='Hotspot name')
    parser.add_argument('--password', help='Hotspot password')
    parser.add_argument('--authentication', choices=[t.value for t in AuthType],
                        default='wpa-psk', help='Authentication type')
    parser.add_argument('--encryption', choices=[t.value for t in EncryptionType],
                        default='aes', help='Encryption type')
    parser.add_argument('--channel', type=int,
                        default=11, help='Channel number')
    parser.add_argument('--max-clients', type=int, default=10,
                        help='Maximum number of clients')
    parser.add_argument(
        '--band', choices=['bg', 'a', 'ac'], default='bg', help='WiFi band')
    parser.add_argument('--hidden', action='store_true', help='Hidden network')
    parser.add_argument(
        '--profile', help='Profile name for save/load operations')

    args = parser.parse_args()

    manager = HotspotManager()

    try:
        if args.action == 'start':
            config = HotspotConfig(
                name=args.name,
                password=args.password,
                authentication=AuthType(args.authentication),
                encryption=EncryptionType(args.encryption),
                channel=args.channel,
                max_clients=args.max_clients,
                band=args.band,
                hidden=args.hidden
            )
            await manager.start(config)

        elif args.action == 'stop':
            await manager.stop()

        elif args.action == 'status':
            status = await manager.status()
            print(json.dumps(status, indent=2))

        elif args.action == 'list-clients':
            clients = await manager.list_clients()
            print(json.dumps(clients, indent=2))

        elif args.action == 'save-profile':
            if not args.profile:
                raise ValueError(
                    "Profile name is required for save-profile action")
            config = HotspotConfig(
                name=args.name,
                authentication=AuthType(args.authentication),
                encryption=EncryptionType(args.encryption),
                channel=args.channel,
                max_clients=args.max_clients,
                band=args.band,
                hidden=args.hidden
            )
            manager.save_profile(args.profile, config)

        elif args.action == 'load-profile':
            if not args.profile:
                raise ValueError(
                    "Profile name is required for load-profile action")
            config = manager.load_profile(args.profile)
            print(json.dumps(config.__dict__, indent=2))

    except Exception as e:
        logger.error(f"Error: {str(e)}")
        sys.exit(1)

if __name__ == "__main__":
    asyncio.run(main())
